fixing warning for desktop builds

+++ Core/SVGPathElement.m
@@ -431,7 +431,7 @@ digit:
 
 - (CGFloat) readCoordinate:(NSScanner*)scanner
 {
-    CGFloat f;
+    float f;
     BOOL ok;
     ok = [scanner scanFloat:&f];
     NSAssert(ok, @"invalid coord");
